requirements.txt
bleach==1.4.2
chardet==2.3.0
Django==1.9.2
django-braces==1.8.1
django-debug-toolbar==1.4
django-filter==0.12.0
django-mptt==0.8.0
django-rest-swagger==0.3.4
djangorestframework==3.3.2
future==0.15.2
html5lib==0.9999999
mysqlclient==1.3.6
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
sqlparse==0.1.18
